Centralise printable window flag code.
Rather than hardcode where the printable window flags need to be used (such
as on the status-line and choose-window list), introduce a
printable_window_flags() function to return a list of flags.

As a change of behaviour, now returns *all* window flags on a given winlink
rather than a single flag.